New Datamanagment for Photos
New handling of the data for phots:

String "Fotobearbeitung" isgnals the mode
Raw Byte Data comes in (no conversion)
The Raw data is written into a string and converted back into a Byte
Array.

The Byte Array should fill a Bitmap.
But the .FromByteArray Function is not documentent and throws outofRange
Exceptions